pyproject.toml
[tool.poetry]
name = "mars"
version = "0.1.0"
description = "Convert image file extensions from one type to another"
authors = ["binhnguyen00 <jackjack2000.kahp@gmail.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
pillow = "^10.4.0"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
pyyaml = "^6.0.2"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
imageio = "^2.35.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
